+++ Sunrise.Server.Tests/Core/DatabaseFixture.cs
@@ -1,4 +1,5 @@
 ﻿using Sunrise.Server.Application;
+using Sunrise.Server.Tests.Core.Utils;
 
 namespace Sunrise.Server.Tests.Core;
 
@@ -6,6 +7,9 @@ public class DatabaseFixture : EnvironmentFixture, IDisposable
 {
     public new void Dispose()
     {
+        if (!Configuration.DataPath.IsDevelopmentFile())
+            throw new InvalidOperationException("Data path is not a development directory. Are you trying to delete production data?");
+
         if (Directory.Exists(Configuration.DataPath))
             Directory.Delete(Path.Combine(Configuration.DataPath), true);
